feat: use /status for swap execution handling
