--- src/executionFiles/exchanges/oneinch.unit.spec.ts
@@ -1,110 +1,0 @@
-import {
-  TransactionReceipt,
-  TransactionResponse,
-} from '@ethersproject/providers'
-import { BigNumber } from 'ethers'
-import { ParsedReceipt } from '../../types'
-import { oneinch } from './oneinch'
-
-async function getAndTestTransaction(
-  tx: TransactionResponse,
-  receipt: TransactionReceipt,
-  expected: ParsedReceipt
-) {
-  const parsed = await oneinch.parseReceipt(tx, receipt)
-  const needed = {
-    fromAmount: parsed.fromAmount,
-    toAmount: parsed.toAmount,
-    gasUsed: parsed.gasUsed,
-    gasPrice: parsed.gasPrice,
-    gasFee: parsed.gasFee,
-  }
-
-  expect(needed).toEqual(expected)
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function reviveBigNumberFromJson(key: any, value: any) {
-  let result = value
-  if (
-    typeof value === 'object' &&
-    value !== null &&
-    value.hasOwnProperty('type')
-  ) {
-    switch (value.type) {
-      case 'BigNumber':
-        result = BigNumber.from(value.hex)
-    }
-  }
-  return result
-}
-
-/* eslint-disable max-len */
-describe('oneinch', () => {
-  describe('parse receipt', () => {
-    describe('POL', () => {
-      it('gas to token', async () => {
-        // https://polygonscan.com/tx/0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396
-        const tx = JSON.parse(
-          '{"hash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","type":0,"accessList":null,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3","blockNumber":23066100,"transactionIndex":8,"confirmations":310317,"from":"0x0a8fCfeFC82e368045FF30835C4B4201B84200b5","gasPrice":{"type":"BigNumber","hex":"0x174876e800"},"gasLimit":{"type":"BigNumber","hex":"0x062df4"},"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","value":{"type":"BigNumber","hex":"0x03e996576ba8efdb1e"},"nonce":34,"data":"0x7c02520000000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000180000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa8417400000000000000000000000011431a89893025d2a48dca4eddc396f8c81171870000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000000000000000ae0bd9c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000860000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000004e000000000000000000000000000000000000000000000000000000000000005c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000144b3af37c000000000000000000000000000000000000000000000000000000000000000804000000000000000000000000000000000000000000000000000000000000004000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee00000000000000000000000000000640000000000000000000000000000006400000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf12700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000004d0e30db00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001a4b3af37c0000000000000000000000000000000000000000000000000000000000000008080000000000000000000000000000000000000000000000000000000000000440000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270000000000000000000000000000006400000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064d1660f990000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000010dd6d8a29d489bede472cc1b22dc695c144c5c70000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000010dd6d8a29d489bede472cc1b22dc695c144c5c70000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000024bd6015b400000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001a4b3af37c0000000000000000000000000000000000000000000000000000000000000008080000000000000000000000000000000000000000000000000000000000000440000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000000000010000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064d1660f990000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa841740000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b500000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000cfee7c08","r":"0xa845e01c9ee097d57023cb9750ea6857cc830ff03ed30148d6a0a5e7048d3ef1","s":"0x71dce09afb132ef5bb65529374b8cad7b362a69b496c6b1c5fd1e18e2cbf48eb","v":309,"creates":null,"chainId":137}',
-          reviveBigNumberFromJson
-        )
-        const receipt = JSON.parse(
-          '{"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","from":"0x0a8fCfeFC82e368045FF30835C4B4201B84200b5","contractAddress":null,"transactionIndex":8,"gasUsed":{"type":"BigNumber","hex":"0x0333cb"},"logsBloom":"0x0000000000000000000000000000000000000008000000000000000028000000000100000000048000001010000101000000800000000000000000000000000000000000000040000000000800000080000004008000000000010000800002000000000000000000000100000000000000000000000000c580000010000000000001000000000000000000000000000004000001000000000000000000000000200000000000000000000020020000004040000000000000000000000000004000100002000000000001000000001000080000000001800102108080000000000000008000000000000000000000000001000000010080400000000000100800","blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3","transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","logs":[{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0000000000000000000000000000000000001010","topics":["0xe6497e3ee548a3372136af2fcb0696db31fc6cf20260707645068bd3fe97f3c4","0x0000000000000000000000000000000000000000000000000000000000001010","0x0000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5","0x00000000000000000000000011111112542d85b3ef69ae05771c2dccff4faa26"],"data":"0x000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000003f6e72d67e836bb1e000000000000000000000000000000000000000000000002d5f43666538467580000000000000000000000000000000000000000000000000d50d5fc3f46e000000000000000000000000000000000000000000000000006bf8a8dd1fc744276","logIndex":43,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0000000000000000000000000000000000001010","topics":["0xe6497e3ee548a3372136af2fcb0696db31fc6cf20260707645068bd3fe97f3c4","0x0000000000000000000000000000000000000000000000000000000000001010","0x00000000000000000000000011111112542d85b3ef69ae05771c2dccff4faa26","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000006bf8a8dd1fc7442760000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002d5f4366653846758000000000000000000000000000000000000000000000003e996576ba8efdb1e","logIndex":44,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0000000000000000000000000000000000001010","topics":["0xe6497e3ee548a3372136af2fcb0696db31fc6cf20260707645068bd3fe97f3c4","0x0000000000000000000000000000000000000000000000000000000000001010","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270"],"data":"0x000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000003e996576ba8efdb1e0000000000000000000000000000000000000000010c6496a35c93434c6b2d8600000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010c649a8cf2eaaef55b08a4","logIndex":45,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270","topics":["0xe1fffcc4923d04b559f4d29a8bfc6cda04eb5b0d3c460751c2402c5c5cc9109c","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000003e996576ba8efdb1e","logIndex":46,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x00000000000000000000000010dd6d8a29d489bede472cc1b22dc695c144c5c7"],"data":"0x000000000000000000000000000000000000000000000003e996576ba8efdb1e","logIndex":47,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000010dd6d8a29d489bede472cc1b22dc695c144c5c7","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000000000000000b36dd64","logIndex":48,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x10Dd6d8A29D489BEDE472CC1b22dc695c144c5c7","topics":["0xc2c0245e056d5fb095f04cd6373bc770802ebd1e6c918eb78fdef843cdb37b0f"],"data":"0x0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf12700000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000000000000000b36dd6400000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","logIndex":49,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x0000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5"],"data":"0x000000000000000000000000000000000000000000000000000000000b36dd64","logIndex":50,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","topics":["0xd6d4f5681c246c9f42c203e287975af1601f8df8035a9251f79aab5c8f09e2f8"],"data":"0x0000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa841740000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5000000000000000000000000000000000000000000000003e996576ba8efdb1e000000000000000000000000000000000000000000000000000000000b36dd64","logIndex":51,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"},{"transactionIndex":8,"blockNumber":23066100,"transactionHash":"0x88189edfbfaf47a1c8ae0b22680ff8385380cbdf7d8cf4e35f6c0337e5d23396","address":"0x0000000000000000000000000000000000001010","topics":["0x4dfe1bbbcf077ddc3e01291eea2d5c70c2b422b415d95645b9adcfd678cb1d63","0x0000000000000000000000000000000000000000000000000000000000001010","0x0000000000000000000000000a8fcfefc82e368045ff30835c4b4201b84200b5","0x000000000000000000000000de8da1ee512529b6c61fe7c769affc160308dea2"],"data":"0x000000000000000000000000000000000000000000000000004a8f4ade81f800000000000000000000000000000000000000000000000003f7770e1f5053db1e00000000000000000000000000000000000000000000027ebe3f24787dcf3a0d000000000000000000000000000000000000000000000003f72c7ed471d1e31e00000000000000000000000000000000000000000000027ebe89b3c35c51320d","logIndex":52,"blockHash":"0x46f2bc4e4527390b48d76cf9e7da9ec278c7e75b818ac1a9c72f433772a884b3"}],"blockNumber":23066100,"confirmations":310317,"cumulativeGasUsed":{"type":"BigNumber","hex":"0x19d4de"},"effectiveGasPrice":{"type":"BigNumber","hex":"0x174876e800"},"status":1,"type":0,"byzantium":true}',
-          reviveBigNumberFromJson
-        )
-        const expected = {
-          fromAmount: '72171968998379871006',
-          toAmount: '188144996',
-          gasUsed: '209867',
-          gasPrice: '100000000000',
-          gasFee: '20986700000000000',
-        }
-
-        await getAndTestTransaction(tx, receipt, expected)
-      })
-
-      it('token to gas', async () => {
-        // https://polygonscan.com/tx/0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72
-        const txString =
-          '{"hash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","type":0,"accessList":null,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498","blockNumber":23066254,"transactionIndex":17,"confirmations":310163,"from":"0x15177a71AF4ddE1e190c6a183a8AFCf39ae2224A","gasPrice":{"type":"BigNumber","hex":"0x1336f979a5"},"gasLimit":{"type":"BigNumber","hex":"0x0ff54b"},"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","value":{"type":"BigNumber","hex":"0x00"},"nonce":10827,"data":"0x6c4a483e00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001a0000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a00000000000000000000000000000000000000000000000000000000d09dc30000000000000000000000000000000000000000000000004872a17558fa480000000000000000000000000000000000000000000000000000000000000000000d00000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000019a0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000078000000000000000000000000000000000000000000000000000000000000013e0000000000000000000000000000000000000000000000000000000000000170000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000644b3af37c000000000000000000000000000000000000000000000000000000000000000808000000000000000000000000000000000000000000000000000000000000024000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000000000500000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000005045636885000000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000001000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a00000000000000000000000011431a89893025d2a48dca4eddc396f8c81171870000000000000000000000000000000000000000000000121e2412496a4d5fe90000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000002a0800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001a4b3af37c000000000000000000000000000000000000000000000000000000000000000808000000000000000000000000000000000000000000000000000000000000044000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000000000f0000000000000000000000000000003c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064d1660f99000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f00000000000000000000000065d43b64e3b31965cd5ea367d4c2b94c0308479700000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000a4b757fed600000000000000000000000065d43b64e3b31965cd5ea367d4c2b94c03084797000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf12700000000000000000001e848011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000ba4b3af37c000000000000000000000000000000000000000000000000000000000000000808000000000000000000000000000000000000000000000000000000000000024000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000000000f0000000000000000000000000000000f00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000a645636885000000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000001000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000036599834ff9575586c0000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000002c000000000000000000000000000000000000000000000000000000000000003a0800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001a4b3af37c000000000000000000000000000000000000000000000000000000000000000808000000000000000000000000000000000000000000000000000000000000044000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000000002d0000000000000000000000000000002d00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064d1660f99000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f00000000000000000000000038c56f05a7b4e4dc79eac6c264829c7b21fbf7a80000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000038c56f05a7b4e4dc79eac6c264829c7b21fbf7a80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000024bd6015b400000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000050483f1291f000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000004a00000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000000000030000000000000000000000000000000300000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000016080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064eb5625d90000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000800000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001e452bbbe2900000000000000000000000000000000000000000000000000000000000000e000000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000000000000000000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000061cc6f680297e37f1873d2dab4487aa67cd56b58e2f2787500010000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa841740000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000280000000000000000000000000000000000000000000000000000000000000448000000000000000000000000000000000000000000000000000000000000164000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000002647f8fe7a00000000000000000000000000000000000000000000000000000000000000080800000000000000000000000000000000000000000000000000000000000004400000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000000000000000000000000000000000000000001e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000a4059712240000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000100000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004470bdb9470000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf12700000000000000000000000000000000000000000000000487997401914103d0d0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001a4b3af37c0000000000000000000000000000000000000000000000000000000000000008080000000000000000000000000000000000000000000000000000000000000440000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270000000000000000000000000000000010000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064d1660f990000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","r":"0x3f8753bc5aa7c1b075bcc8f7d4ef3cb045241e30cc010d8d8a6166ff792dead6","s":"0x1278048fc6ca47b0d5722cddc38f96a73764ba430c3b46c214270f27dc80dc31","v":310,"creates":null,"chainId":137}'
-        const tx = JSON.parse(txString, reviveBigNumberFromJson)
-        const receiptString =
-          '{"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","from":"0x15177a71AF4ddE1e190c6a183a8AFCf39ae2224A","contractAddress":null,"transactionIndex":17,"gasUsed":{"type":"BigNumber","hex":"0x051844"},"logsBloom":"0x00200000000000000000000080004000000000080000000000000000280000001000000200000400000110000001000108008000000000000000000000200000000000000000400000200008018020a00200040000000000000102000000020000000000020002000001110800008800000000400000004180000010000000000001000000080000000000000000000000000000000010082000004000000000a20000000000050000100020080000004040000000000000000180020000004400000002000000000181000000001000482000000001001000100080001020000010008000000000000000000000000000800000014080100000040000100800","blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498","transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","logs":[{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0xc2132D05D31c914a87C6611C10748AEb04B58e8F","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x00000000000000000000000000000000000000000000000000000000d09dc300","logIndex":208,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0xc2132D05D31c914a87C6611C10748AEb04B58e8F","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a","0x00000000000000000000000011111112542d85b3ef69ae05771c2dccff4faa26"],"data":"0x7ffffffffffffffffffffffffffffffffffffffffffffffffffff5f300ba572a","logIndex":209,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0xc2132D05D31c914a87C6611C10748AEb04B58e8F","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x00000000000000000000000065d43b64e3b31965cd5ea367d4c2b94c03084797"],"data":"0x00000000000000000000000000000000000000000000000000000000342770c0","logIndex":210,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000065d43b64e3b31965cd5ea367d4c2b94c03084797","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x0000000000000000000000000000000000000000000000121e9ad1d8d291063c","logIndex":211,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x65D43B64E3B31965Cd5EA367D4c2b94c03084797","topics":["0x1c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1"],"data":"0x00000000000000000000000000000000000000000000808ec81cf995c23d2081000000000000000000000000000000000000000000000000000001717ed32710","logIndex":212,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x65D43B64E3B31965Cd5EA367D4c2b94c03084797","topics":["0xd78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d822","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000342770c00000000000000000000000000000000000000000000000121e9ad1d8d291063c0000000000000000000000000000000000000000000000000000000000000000","logIndex":213,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0xc2132D05D31c914a87C6611C10748AEb04B58e8F","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x00000000000000000000000038c56f05a7b4e4dc79eac6c264829c7b21fbf7a8"],"data":"0x000000000000000000000000000000000000000000000000000000009c765240","logIndex":214,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000038c56f05a7b4e4dc79eac6c264829c7b21fbf7a8","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000000000000009c78de7e","logIndex":215,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x38C56f05a7b4E4dC79Eac6C264829C7b21fBf7A8","topics":["0xc2c0245e056d5fb095f04cd6373bc770802ebd1e6c918eb78fdef843cdb37b0f"],"data":"0x000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000000000000000000000000000000000009c765240000000000000000000000000000000000000000000000000000000009c78de7e00000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","logIndex":216,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x000000000000000000000000000000000000000000000000000000009c78de7e","logIndex":217,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0xBA12222222228d8Ba445958a75a0704d566BF2C8","topics":["0x2170c741c41531aec20e7c107c24eecfdd15e69c9bb0a8dd37b1840b9e0b207b","0x0297e37f1873d2dab4487aa67cd56b58e2f27875000100000000000000000002","0x0000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174","0x0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf1270"],"data":"0x000000000000000000000000000000000000000000000000000000009c78de7e0000000000000000000000000000000000000000000000365afc6e40417f36d1","logIndex":218,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x000000000000000000000000000000000000000000000000000000009c78de7e","logIndex":219,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x0000000000000000000000000000000000000000000000000000000000000000","logIndex":220,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x0000000000000000000000000000000000000000000000365afc6e40417f36d1","logIndex":221,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a"],"data":"0x0000000000000000000000000000000000000000000000487997401914103d0d","logIndex":222,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","topics":["0xd6d4f5681c246c9f42c203e287975af1601f8df8035a9251f79aab5c8f09e2f8"],"data":"0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a000000000000000000000000c2132d05d31c914a87c6611c10748aeb04b58e8f0000000000000000000000000d500b1d8e8ef31e21c99d1db9a6444d3adf127000000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a00000000000000000000000000000000000000000000000000000000d09dc3000000000000000000000000000000000000000000000000487997401914103d0d","logIndex":223,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0000000000004946c0e9F43F4Dee607b0eF1fA1c","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a","0x0000000000000000000000000000000000000000000000000000000000000000"],"data":"0x000000000000000000000000000000000000000000000000000000000000000d","logIndex":224,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0000000000004946c0e9F43F4Dee607b0eF1fA1c","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a","0x00000000000000000000000011111112542d85b3ef69ae05771c2dccff4faa26"],"data":"0x7ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe171c","logIndex":225,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"},{"transactionIndex":17,"blockNumber":23066254,"transactionHash":"0x487aabcdbb3a0334328e300988e5f0a4c86e0091da33136f99f71616c090ed72","address":"0x0000000000000000000000000000000000001010","topics":["0x4dfe1bbbcf077ddc3e01291eea2d5c70c2b422b415d95645b9adcfd678cb1d63","0x0000000000000000000000000000000000000000000000000000000000001010","0x00000000000000000000000015177a71af4dde1e190c6a183a8afcf39ae2224a","0x000000000000000000000000b79fad4ca981472442f53d16365fdf0305ffd8e9"],"data":"0x0000000000000000000000000000000000000000000000000061e5215de4c7d400000000000000000000000000000000000000000000000090cbc8966d7aa5f400000000000000000000000000000000000000000000dbcce132d03c6ba4d6c10000000000000000000000000000000000000000000000009069e3750f95de2000000000000000000000000000000000000000000000dbcce194b55dc9899e95","logIndex":226,"blockHash":"0x82b14b424ca5f158adb62f1fd3ca8de9bbdbb911dc58e6762ba868fa2fb6f498"}],"blockNumber":23066254,"confirmations":310163,"cumulativeGasUsed":{"type":"BigNumber","hex":"0x5ee8e1"},"effectiveGasPrice":{"type":"BigNumber","hex":"0x1336f979a5"},"status":1,"type":0,"byzantium":true}'
-        const receipt = JSON.parse(receiptString, reviveBigNumberFromJson)
-
-        const expected = {
-          fromAmount: '3500000000',
-          toAmount: '1336927115383615274253',
-          gasUsed: '333892',
-          gasPrice: '82526697893',
-          gasFee: '27555004212889556',
-        }
-
-        await getAndTestTransaction(tx, receipt, expected)
-      })
-
-      it('token to token', async () => {
-        // https://polygonscan.com/tx/0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da
-        const tx = JSON.parse(
-          '{"hash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","type":0,"accessList":null,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0","blockNumber":23080384,"transactionIndex":44,"confirmations":296033,"from":"0x2A63a682B34c92EA39083402aCdfB4f77c2950b3","gasPrice":{"type":"BigNumber","hex":"0x174876e800"},"gasLimit":{"type":"BigNumber","hex":"0x042541"},"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","value":{"type":"BigNumber","hex":"0x00"},"nonce":113691,"data":"0x7c02520000000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000001800000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa8417400000000000000000000000053e0bca35ec356bd5dddfebbd1fc0fd03fabad3900000000000000000000000011431a89893025d2a48dca4eddc396f8c81171870000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3000000000000000000000000000000000000000000000000000000001dcd65000000000000000000000000000000000000000000000000014f5e4a38ca0c072500000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006200000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000208000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000050483f1291f000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000004a00000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000000006400000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000016080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000064eb5625d90000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000800000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000001e452bbbe2900000000000000000000000000000000000000000000000000000000000000e000000000000000000000000011431a89893025d2a48dca4eddc396f8c811718700000000000000000000000000000000000000000000000000000000000000000000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000061cce88a36128d5436d2d70cab39c9af9cce146c38554ff000010000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa8417400000000000000000000000053e0bca35ec356bd5dddfebbd1fc0fd03fabad39000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000028000000000000000000000000000000000000000000000000000000000000044800000000000000000000000000000000000000000000000000000000000016400000000000000000000000000000000000000000000000000000000cfee7c08","r":"0x9f74ec5da7af2b310ea7942616048423accca22c1c57f3ee1f18a5153592baa0","s":"0x28abbed95b4a337c5200b426c3bf57dcf10d7952e89b53ad873217204463863e","v":309,"creates":null,"chainId":137}',
-          reviveBigNumberFromJson
-        )
-        const receipt = JSON.parse(
-          '{"to":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","from":"0x2A63a682B34c92EA39083402aCdfB4f77c2950b3","contractAddress":null,"transactionIndex":44,"gasUsed":{"type":"BigNumber","hex":"0x02a2cf"},"logsBloom":"0x00000000000000000000000000004000000000080020000000000000280000000000000000000400001120000001000000008000000002000000000000200000000004000000000000000008000020800000000000000000000108000000000000000000000000000001100000000400000000400004004180000010000000000001000000000000800000000000000000000000000000002000000000000000220000000000010000000000000000004000000000000000000100000000004400000002000000000101000000001000082000000000000000100080001000082010008000002000000000000000000000000000010080000100000000110000","blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0","transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","logs":[{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x0000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187"],"data":"0x000000000000000000000000000000000000000000000000000000001dcd6500","logIndex":251,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x0000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3","0x00000000000000000000000011111112542d85b3ef69ae05771c2dccff4faa26"],"data":"0xfffffffffffffffffffffffffffffffffffffffffffffffffffffb2355c09d56","logIndex":252,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x000000000000000000000000000000000000000000000000000000001dcd6500","logIndex":253,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0xBA12222222228d8Ba445958a75a0704d566BF2C8","topics":["0x2170c741c41531aec20e7c107c24eecfdd15e69c9bb0a8dd37b1840b9e0b207b","0x36128d5436d2d70cab39c9af9cce146c38554ff0000100000000000000000008","0x0000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa84174","0x00000000000000000000000053e0bca35ec356bd5dddfebbd1fc0fd03fabad39"],"data":"0x000000000000000000000000000000000000000000000000000000001dcd65000000000000000000000000000000000000000000000000014f99efc9f5d13922","logIndex":254,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x000000000000000000000000000000000000000000000000000000001dcd6500","logIndex":255,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x00000000000000000000000011431a89893025d2a48dca4eddc396f8c8117187","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8"],"data":"0x0000000000000000000000000000000000000000000000000000000000000000","logIndex":256,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39","topics":["0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef","0x000000000000000000000000ba12222222228d8ba445958a75a0704d566bf2c8","0x0000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3"],"data":"0x0000000000000000000000000000000000000000000000014f99efc9f5d13922","logIndex":257,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x11111112542D85B3EF69AE05771c2dCCff4fAa26","topics":["0xd6d4f5681c246c9f42c203e287975af1601f8df8035a9251f79aab5c8f09e2f8"],"data":"0x0000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b30000000000000000000000002791bca1f2de4661ed88a30c99a7a9449aa8417400000000000000000000000053e0bca35ec356bd5dddfebbd1fc0fd03fabad390000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3000000000000000000000000000000000000000000000000000000001dcd65000000000000000000000000000000000000000000000000014f99efc9f5d13922","logIndex":258,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"},{"transactionIndex":44,"blockNumber":23080384,"transactionHash":"0x129980866efb775cdeeb2b5f5bf65adc04d65ffc0b8caa54902bf856dac1b1da","address":"0x0000000000000000000000000000000000001010","topics":["0x4dfe1bbbcf077ddc3e01291eea2d5c70c2b422b415d95645b9adcfd678cb1d63","0x0000000000000000000000000000000000000000000000000000000000001010","0x0000000000000000000000002a63a682b34c92ea39083402acdfb4f77c2950b3","0x0000000000000000000000007b5000af8ab69fd59eb0d4f5762bff57c9c04385"],"data":"0x000000000000000000000000000000000000000000000000003d5f9ca6f5980000000000000000000000000000000000000000000000002c3b9253bed0bf267d000000000000000000000000000000000000000000000c24da52a393a1f39d8d00000000000000000000000000000000000000000000002c3b54f42229c98e7d000000000000000000000000000000000000000000000c24da90033048e9358d","logIndex":259,"blockHash":"0xdec87ec0b811b9b808fbc02c7a411f8d7f4ffeb0e90d4a2a1759e672316246a0"}],"blockNumber":23080384,"confirmations":296033,"cumulativeGasUsed":{"type":"BigNumber","hex":"0x8b43c3"},"effectiveGasPrice":{"type":"BigNumber","hex":"0x174876e800"},"status":1,"type":0,"byzantium":true}',
-          reviveBigNumberFromJson
-        )
-
-        const expected = {
-          fromAmount: '500000000',
-          toAmount: '24182623324834183458',
-          gasUsed: '172751',
-          gasPrice: '100000000000',
-          gasFee: '17275100000000000',
-        }
-
-        await getAndTestTransaction(tx, receipt, expected)
-      })
-    })
-  })
-})
